requirements.txt
Django==4.0.2
factory-boy==3.2.1
flake8==4.0.1
flake8-docstrings==1.6.0
pytest==7.0.1
pytest-django==4.5.2
weasyprint==54.1
django-imagekit==4.1.0
django-ckeditor==6.2.0
